"Modified dengue_diagnostics.py to use render functions from dengue_envs/envs/dengue_diagnostics_render.py"
